backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
ibapi>=9.81.1
httpx>=0.27.0
supabase>=2.3.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
APScheduler>=3.10.0
python-telegram-bot>=21.0
numpy>=1.26.0
pandas>=2.1.0
